+++ admin/index.php
@@ -6,3 +6,54 @@
  * Time: 12:20 AM
  * To change this template use File | Settings | File Templates.
  */
+error_reporting(E_ALL | E_STRICT);
+ini_set('display_errors',1);
+
+session_start();
+if(empty($_SESSION['is_logged'])) {
+    header( 'Location: login.php' ) ;
+}
+
+
+require_once '../config.php';
+require_once ROOT_PATH.DS.'includes'.DS.'Event.php';
+
+?>
+<!DOCTYPE html>
+<html>
+<head>
+    <link rel="stylesheet" type="text/css" href="main.css"/>
+
+    <script type="text/javascript" src="http://code.jquery.com/jquery-2.0.0.min.js"></script>
+    <script type="text/javascript" src="plugins.js"></script>
+    <script type="text/javascript" src="ace_editor/src-noconflict/ace.js"></script>
+</head>
+<body>
+<ul class="sidebar">
+    <li>
+        <a href="?l=addpost">Add A RawText Post</a>
+    </li>
+    <li>
+        <a href="?l=managepost">Manage Posts</a>
+    </li>
+    <li>
+        <a href="?l=logger">View Log</a>
+    </li>
+</ul>
+<div class="content">
+<?php
+try
+{
+    if(!empty($_GET['l']))
+    {
+        require_once ROOT_PATH.DS.'admin'.DS.$_GET['l'].'.php';
+    }
+}
+catch(Exception $e)
+{
+    echo $e->getMessage();
+}
+?>
+</div>
+</body>
+</html>