+++ admin/logger.php
@@ -4,23 +4,18 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-error_reporting(E_ALL | E_STRICT);
-ini_set('display_errors',1);
 
-require_once '../config.php';
 require_once ROOT_PATH.DS.'includes'.DS.'Event.php';
 
 Event::cleanup();
 $arr = Event::read();
 
 ?>
-
-<html>
-    <head></head>
-    <body>
+<h1>View Log</h1>
+<hr />
         <table style="width:100%;">
 <?php
-foreach($arr as $row) {
+foreach($arr as $row):
     echo '<tr>'.
          '<td style="width:10%">'.$row['id'].'</td>'.
          '<td style="width:10%">'.Event::getBotName($GLOBALS['BOT_CONFIG'], $row['bot_id']).'</td>'.
@@ -28,8 +23,6 @@ foreach($arr as $row) {
          '<td style="width:60%">'.$row['message'].'</td>'.
          '<td style="width:10%">'.$row['time'].'</td>'.
          '</tr>';
-}
+endforeach;
 ?>        
        </table>
-    </body>
-</html>