+++ includes/Post_Collection.php
@@ -7,83 +7,75 @@
 require_once ROOT_PATH.DS.'includes'.DS.'Post.php';
 
 class Post_Collection extends Post_Resource
-{    
+{
+    protected $_where;
+
     function __construct() {
         parent::__construct();
+        $this->_where = array();
     }
-    
-    public function loadDefault($offset = 0, $limit = 10) 
-    {
-        $sql = "SELECT * FROM ".POSTS_TBL." ORDER BY time DESC LIMIT :offset, :limit";
-        $sth = $this->db_res->prepare($sql);
-        $sth->execute(
-                array(
-                    ':offset' => $offset*$limit,
-                    ':limit'  => $limit
-                )); 
-        $rows = $sth->fetchAll();
-        
-        return $this->_fetchRows($rows);        
-    }
-    
-    public function loadByTime($from,$to,$offset = 0,$limit = 10) 
-    {
-        $sql = "SELECT * FROM ".POSTS_TBL." WHERE time >= :from AND time <= :to ORDER BY time DESC LIMIT :offset, :limit";
-        $sth = $this->db_res->prepare($sql);
-        $sth->execute(
-                array(
-                    ':from' => $from, 
-                    ':to' => $to,
-                    ':offset' => $offset*$limit,
-                    ':limit'  => $limit
-                ));
 
-        $rows = $sth->fetchAll();
-        
-        return $this->_fetchRows($rows);
+    function addWhere($name,$op,$value,$isOr = false,$name_postfix = '') {
+        $temp = new stdClass();
+        $temp->isOr = $isOr;
+        $temp->name = $name;
+        $temp->name_postfix = $name_postfix;
+        $temp->op = $op;
+        $temp->value = $value;
+        $this->_where[$name.$name_postfix] = $temp;
     }
     
-    public function loadByProvider($id, $offset = 0,$limit = 10) 
-    {
-        $sql = "SELECT * FROM ".POSTS_TBL." WHERE provider_id = :provider_id ORDER BY time DESC LIMIT :offset, :limit";
-        $sth = $this->db_res->prepare($sql);
-        $sth->execute(
-                array(
-                    ':provider_id' => $provider_id,
-                    ':offset' => $offset*$limit,
-                    ':limit'  => $limit
-                )); 
-        $rows = $sth->fetchAll();
-        
-        return $this->_fetchRows($rows);         
-    } 
-    
-    public function loadByQuery($provider,$query,$from = null,$to = null,$offset = 0, $limit = 10) 
+    public function loadByQuery($offset = 0, $limit = 10)
     {
         
         $sql = "SELECT DISTINCT * FROM ".POSTS_TBL." ";
-        $hasWhere = false;       
-        
-        if(!empty($provider)) 
+
+        $isFirst = true;
+        foreach($this->_where as $val)
+        {
+            if($isFirst) {
+                $sql .= "WHERE {$val->name} {$val->op} :{$val->name}{$val->name_postfix} ";
+                $isFirst = false;
+            }
+            else
+            {
+                if($val->isOr)
+                {
+                    $sql .= "OR {$val->name} {$val->op} :{$val->name}{$val->name_postfix} ";
+                }
+                else
+                {
+                    $sql .= "AND {$val->name} {$val->op} :{$val->name}{$val->name_postfix} ";
+                }
+
+            }
+
+        }
+
+        /*
+        $hasWhere = false;
+
+        if(!empty($provider))
         {
             $sql .= "WHERE provider_id = :provider_id ";
             $hasWhere = true;
         }
-        
+
         $tokens = array();
         
         if(!empty($query)) 
         {
             $tokens = explode(" ",$query);
             if(!$hasWhere)
-            {             
+            {
                 $sql .= "WHERE ";
+                $hasWhere = true;
             }
             else
             {
                 $sql .= "AND ";
             }
-            
+
             $first = true;
 
             foreach($tokens as $idx => $val) 
@@ -104,8 +96,9 @@ class Post_Collection extends Post_Resource
         if(!empty($from) || !empty($to))
         {
             if(!$hasWhere)
-            {             
+            {
                 $sql .= "WHERE ";
+                $hasWhere = true;
             }
             else
             {
@@ -125,40 +118,23 @@ class Post_Collection extends Post_Resource
                 $sql .= 'time <= :to ';
             }        
         }
+        */
         
         $sql .= "ORDER BY time DESC LIMIT :offset, :limit";       
-        
+
+
         $sth = $this->db_res->prepare($sql);
-        
-        if(!empty($provider)) {
-            $sth->bindParam(':provider_id',$provider);
+
+        foreach($this->_where as $val) {
+            $sth->bindValue(':'.$val->name.$val->name_postfix,$val->value);
         }
-        
-        
-        if(!empty($tokens)) {
-            foreach($tokens as $idx => $val) {
-                $sth->bindValue(':qy'.$idx,'%'.$val.'%');
-            }
-        }
-        
-        if(!empty($from)) {
-           $sth->bindValue(':from',$from);
-        }
-       
-        if(!empty($to))
-        {
-           $sth->bindValue(':to',$to);   
-        }
-         
-        $sth->bindValue(':offset',$offset*$limit);
+
+        $sth->bindValue(':offset',$offset);
         $sth->bindValue(':limit',$limit);
-                
-        
+
         $sth->execute();          
         
-        $rows = $sth->fetchAll();       
-        
-        //$rows = $this->db_res->query('SELECT * FROM posts WHERE provider_id = 1 AND  contents LIKE \'%UX%\' ORDER BY time DESC LIMIT 0, 10');
+        $rows = $sth->fetchAll();
         
         return $this->_fetchRows($rows);                 
     }
@@ -177,9 +153,31 @@ class Post_Collection extends Post_Resource
             $post->tags = $row['tags'];
             $post->custom_data = $row['custom_data'];
             $post->time = $row['time'];
+            $post->hidden = $row['hidden'];
+            $post->update_time = $row['update_time'];
             $rc[$post->id] = $post;
         }
         
         return $rc;
+    }
+
+    public function getSize($providerId = 0) {
+        if($providerId)
+        {
+            $sql = "SELECT count(*) FROM ".POSTS_TBL." WHERE provider_id = :provider_id";
+            $sth = $this->db_res->prepare($sql);
+            $sth->bindValue(':provider_id',(int) $providerId);
+        }
+        else
+        {
+            $sql = "SELECT count(*) FROM ".POSTS_TBL;
+            $sth = $this->db_res->prepare($sql);
+        }
+
+        $sth->execute();
+
+        $result = $sth->fetchAll();
+
+        return $result[0][0];
     }
 }